scripts/sct_denoising_onlm.py: fix indexError
Former-commit-id: 0fe66f32480204815fbc64ebf5927013c18a4295

+++ scripts/sct_denoising_onlm.py
@@ -132,7 +132,7 @@ def main(file_to_denoise, param, output_file_name) :
     sct.printv("total time: %s" % (time() - t))
     sct.printv("vol size", den.shape)
 
-    axial_middle = data.shape[2] / 2
+    axial_middle = int(data.shape[2] / 2)
 
     before = data[:, :, axial_middle].T
     after = den[:, :, axial_middle].T
